requirements.txt
Flask>=2.3.0
requests>=2.28.0
pyodbc>=5.0.0
azure-storage-blob>=12.19.0
msal>=1.24.0
gunicorn>=21.0.0
Werkzeug>=2.3.0
